history info: add groups and environments to ouput
