+++ libdnf5-cli/output/transactioninfo.cpp
@@ -22,8 +22,6 @@ along with libdnf.  If not, see <https://www.gnu.org/licenses/>.
 
 #include "fmt/chrono.h"
 
-#include "libdnf5-cli/tty.hpp"
-
 
 namespace libdnf5::cli::output {
 
@@ -44,30 +42,10 @@ void print_transaction_info(libdnf5::transaction::Transaction & transaction) {
     info.add_line("Description", transaction.get_description());
     info.add_line("Comment", transaction.get_comment());
 
-    info.add_line("Packages altered", "");
-
-    std::unique_ptr<libscols_table, decltype(&scols_unref_table)> item_list(scols_new_table(), &scols_unref_table);
-    if (libdnf5::cli::tty::is_interactive()) {
-        scols_table_enable_colors(item_list.get(), 1);
-    }
-
-    // The two spaces indent the table the same way as child lines in KeyValueTable
-    scols_table_new_column(item_list.get(), "  Action", 0, 0);
-    scols_table_new_column(item_list.get(), "Package", 0, 0);
-    scols_table_new_column(item_list.get(), "Reason", 0, 0);
-    scols_table_new_column(item_list.get(), "Repository", 0, 0);
-
-    for (auto & pkg : transaction.get_packages()) {
-        struct libscols_line * ln = scols_table_new_line(item_list.get(), NULL);
-        scols_line_set_data(
-            ln, 0, ("  " + libdnf5::transaction::transaction_item_action_to_string(pkg.get_action())).c_str());
-        scols_line_set_data(ln, 1, pkg.to_string().c_str());
-        scols_line_set_data(ln, 2, libdnf5::transaction::transaction_item_reason_to_string(pkg.get_reason()).c_str());
-        scols_line_set_data(ln, 3, pkg.get_repoid().c_str());
-    }
-
     info.print();
-    scols_print_table(item_list.get());
+    print_transaction_item_table(transaction.get_packages(), "Packages altered:");
+    print_transaction_item_table(transaction.get_comps_groups(), "Groups altered:");
+    print_transaction_item_table(transaction.get_comps_environments(), "Environments altered:");
 }
 
 }  // namespace libdnf5::cli::output
